Improve Reset emulation
Currently, the emulation (for versions of Windows without an ANSI-aware console) fills blank space with 'colour 0' (usually black) and 'resets' to 'colour 7 on colour 0' (usually silver on black). This may upset PowerShell users (or others) who have a different default - for example, Reset does not return them to where they started.

This proposal changes the emulation so that the default colours are those of the console when the emulation starts.

The default colours are captured in module System.Console.ANSI.Windows.Detect and preserved: isANSIEnabled is now a value of type ANSIEnabledStatus (rather than Bool) which has a constructor NotANSIEnabled ConsoleDefaultState. The emulated version of the affected functions (eg hClearScreen and hSetSGR) are then a function of the 'console default state' (as well as the other arguments).

This sees: (1) some of the original Common-Include.hs split out into Common-Include-Enabled.hs or, as applicable, Common-Include-Emulator.hs; and (2) if then else choices in module System.Console.ANSI.Windows become case of choices.