Fix emulation of Erase codes
The current emulation of Erase in Page (ED) and Erase in Line (EL) ANSI codes does not work when the display window is narrower than the console screen buffer (that is, when there is a horizontal scroll bar displayed).

Versions of the Windows console that are ANSI-enabled apply the Erase codes only to the contents of the display window and not other parts of the console screen buffer.

This proposal fixes the emulation of the Erase codes by emulating the current behaviour of the ANSI-enabled Windows consoles.