Fix Data.Typeable in System.Console.ANSI.Windows.Foreign
In the current module System.Console.ANSI.Windows.Foreign, module Data.Typeable is imported only if the GHC version >= 6.12 (base-4.2.0.0, Dec 2009) but type ConsoleException is an instance of class Typeable for all versions of GHC. Also, the import of Data.Typeable is duplicated if the version of the Win32 package is >= 2.5.1. It appears the former (the >= 6.12 constraint) dates from when only function cast (from Data.Typeable) was used, in versions of withHandleToHANDLE implemented for GHC versions >= 6.12.

This commit removes the GHC version >= 6.12 constraint for the import of Data.Typeable; eliminates the duplicated import; and, consistent with Johan Tibell (hindent) style, uses explicit imports for Data.Typeable (namely Typeable and cast) and places the import in alphabetical order in relation to other imports.

In respect of the removal of the GHC version >= 6.12 constraint, the .cabal file specifies base >=4 (GHC 6.10.1, Nov 2008) and Data.Typeable (Typeable, cast) were available in package base-4.0.0.0.