Add support for 256-color palettes
This builds on Olivier Sohn's PR #38, but proposes a different approach to the API. As indicated in the Haddock documentation, the API is used as follows:

`setSGR [ SetPaletteColor $ xterm6LevelRGB 5 2 0 ]`
`setSGR [ SetPaletteColor $ xterm24LevelGray 12 ]`
`setSGR [ SetPaletteColor $ xtermSystem Vivid Green ]`
`setSGR [ SetPaletteColor 15 ]`

A new constructor, `SetPaletteColor ConsoleLayer Word8`, is added to the `SGR` type. This adds to `SGR` the last of the colour extensions to the 'ANSI' standards: indexed colours.

A terminal may not be following xterm's standard protocol for a 256-colour palette. However, 'constructor-like' functions are provided to return indices to the parts of that palette, namely `xtermSystem` (for the 16 system colors), `xterm6LevelRGB` (for the 216 6x6x6 RGB colors) and `xterm24LevelGray` (for the 24 grays).

Changes are made to documentation.