Treat valid non-ConHost stdout as ANSI-enabled
Fixes the remainder of #52. The logic of function enableANSIOut in module System.Console.ANSI.Windows.Detect is changed so that if there is valid standard output handle but the handle is not a ConHost handle, then the handle is assumed to be ANSI-enabled. This conforms the approach to redirected standard output on Windows to be the same as on Unix-like OSs.