Refactor into System.Win32.Compat
The remaining CPP pragmas in module `System.Console.ANSI.Windows.Foreign` are there to provide functions available in later versions of the `Win32` package that did not ship with versions of GHC before GHC 8.2.1. This refactors the code to make that explicit, creating module `System.Win32.Compat` (following the naming convention used in package `base-compat`). The intention is to make the logic of the `ansi-terminal` code more explicit.