Refactor the Windows.Detect module
The detection code in module `System.Console.ANSI.Windows.Detect` is simplified and, with one exception, without changing the algorithm. As a consequence, it is discovered that the desired behaviour does not depend on looking up the `TERM` environment variable and there is no longer a need to rely on the `base-compat` package.

The exception is that, for a ConHost console, an attempt is always made to enable virtual terminal processing, irrespective of the existence and value of the `TERM` environment variable. This now deals with, for example, the case of a Hyper 2 terminal where the underlyng shell is ConHost; Hyper 2 does set `TERM` in those circumstances.

Some identifiers that are not exported as part of the API are renanmed. `ANSIEnabledStatus (ANSIEnabled, NotANSIEnabled)` is renamed `ANSISupport (Native, Emulated)` to reflect better its use. Similarly, `isANSIEnabled` is renamed `aNSISupport`. Those changes flow through to module `System.Console.ANSI.Windows`.

Documentation is updated accordingly.

In passing, the `stack.yaml` resolver is also bumped from LTS 11.0 to 11.7.

The changes have been tested on Windows 10 and legacy Windows (Windows 7) (both ConHost and mintty; and, in the case of ConHost, directly and via the VS Code integrated terminal and Hyper 2); and in the Appveyor build environment.