Cursor pos docs (0- and 1-based) and `getCursorPosition0` added
The 'ANSI' standards are 1-based but the long-settled API for `setCursorColumn` and `setCursorPosition` is 0-based. The 'ANSI' standards determine what is emitted by `reportCursorPosition` and, consequently, the values of `getCursorPosition`. This pull request enhances the Haddock documentation of those families of functions to seek to reduce the chance that users will assume the `set` and `get` functions number rows and columns in the same way.

It also adds 0-based `getCursorPosition0` and deprecates 1-based `getCursorPosition`.